backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
PyJWT==2.11.0
bcrypt==4.1.2
python-dotenv==1.0.0
groq==0.4.1
PyPDF2==3.0.1
python-multipart==0.0.6
reportlab==4.0.9
requests==2.31.0
google-auth==2.26.2
